Move getters for sweetalert2 properties to getters
